% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-file.R
\name{backup_file}
\alias{backup_file}
\title{Create backup of file}
\usage{
backup_file(file)
}
\arguments{
\item{file}{Character. Path to file to backup.}
}
\value{
Path to backup file, invisibly.
}
\description{
Create backup of file
}
