% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-file.R
\name{determine_insert_position}
\alias{determine_insert_position}
\title{Determine where to insert header}
\usage{
determine_insert_position(content)
}
\arguments{
\item{content}{Character vector. File content lines.}
}
\value{
Integer. Position to insert header (0 for beginning of file).
}
\description{
Determine where to insert header
}
\keyword{internal}
