% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{ensure_directory_exists}
\alias{ensure_directory_exists}
\title{Ensure directory exists}
\usage{
ensure_directory_exists(dir)
}
\arguments{
\item{dir}{Character. Path to directory.}
}
\value{
TRUE invisibly if directory exists, otherwise aborts.
}
\description{
Ensure directory exists
}
\keyword{internal}
