% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{ensure_file_exists}
\alias{ensure_file_exists}
\title{Ensure file exists}
\usage{
ensure_file_exists(file)
}
\arguments{
\item{file}{Character. Path to file.}
}
\value{
TRUE invisibly if file exists, otherwise aborts.
}
\description{
Ensure file exists
}
\keyword{internal}
