% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{ensure_valid_action}
\alias{ensure_valid_action}
\title{Ensure action is valid}
\usage{
ensure_valid_action(action)
}
\arguments{
\item{action}{Character. Action to check.}
}
\value{
TRUE invisibly if action is valid, otherwise aborts.
}
\description{
Ensure action is valid
}
\keyword{internal}
