% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{ensure_valid_template}
\alias{ensure_valid_template}
\title{Ensure template is valid}
\usage{
ensure_valid_template(template)
}
\arguments{
\item{template}{Object. Template to check.}
}
\value{
TRUE invisibly if template is valid, otherwise aborts.
}
\description{
Ensure template is valid
}
\keyword{internal}
