% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{extract_years}
\alias{extract_years}
\title{Extract years from a string}
\usage{
extract_years(text)
}
\arguments{
\item{text}{Character. Text to extract years from.}
}
\value{
Numeric vector of years.
}
\description{
Extract years from a string
}
\keyword{internal}
