% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language-support.R
\name{format_header}
\alias{format_header}
\title{Format header according to language's comment style}
\usage{
format_header(content, language)
}
\arguments{
\item{content}{Character. Header content.}

\item{language}{stamp_language object. Language for comment formatting.}
}
\value{
Character. Formatted header.
}
\description{
Format header according to language's comment style
}
\keyword{internal}
