% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_encoding}
\alias{guess_encoding}
\title{Guess file encoding}
\usage{
guess_encoding(content)
}
\arguments{
\item{content}{Raw vector. File content.}
}
\value{
Character. Encoding name.
}
\description{
Guess file encoding
}
\keyword{internal}
