% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_line_ending}
\alias{guess_line_ending}
\title{Guess line endings}
\usage{
guess_line_ending(content)
}
\arguments{
\item{content}{Raw vector. File content.}
}
\value{
Character. Line ending: \code{"\\n"} (LF), \code{"\\r"} (CR), or \code{"\\r\\n"} (CRLF).
}
\description{
Guess line endings
}
\keyword{internal}
