% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{header_field_regex}
\alias{header_field_regex}
\title{Build the header field regular expression}
\usage{
header_field_regex(require_marker = FALSE)
}
\arguments{
\item{require_marker}{Logical. Require a comment marker before the field?}
}
\value{
Character. A perl-compatible regular expression.
}
\description{
Build the header field regular expression
}
\keyword{internal}
