% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{header_file_info}
\alias{header_file_info}
\title{Get file information (encoding, line endings, etc.)}
\usage{
header_file_info(file)
}
\arguments{
\item{file}{Character. Path to file.}
}
\value{
stamp_file_info object.
}
\description{
Get file information (encoding, line endings, etc.)
}
\keyword{internal}
