% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_file}
\alias{is_file}
\title{Check if path is a file}
\usage{
is_file(path)
}
\arguments{
\item{path}{Character. Path to check.}
}
\value{
Logical. TRUE if path is a file.
}
\description{
Check if path is a file
}
