% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language-support.R
\name{language_get}
\alias{language_get}
\title{Get registered language}
\usage{
language_get(name)
}
\arguments{
\item{name}{Character. Language name.}
}
\value{
stamp_language object.
}
\description{
Get registered language
}
