% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-file.R
\name{modify_file}
\alias{modify_file}
\title{Modify file with header}
\usage{
modify_file(file, template, language, ...)
}
\arguments{
\item{file}{Character. Path to file to modify.}

\item{template}{stamp_template object. Template to use for stamping.}

\item{language}{stamp_language object. Detected language for formatting comments.}

\item{...}{Additional arguments passed to render_template.}
}
\value{
TRUE invisibly on success.
}
\description{
Modify file with header
}
\keyword{internal}
