% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-file.R
\name{preview_stamp}
\alias{preview_stamp}
\title{Preview stamp without modifying}
\usage{
preview_stamp(file, template, language, ...)
}
\arguments{
\item{file}{Character. Path to file to preview.}

\item{template}{stamp_template object. Template to use for stamping.}

\item{language}{stamp_language object. Detected language for formatting comments.}

\item{...}{Additional arguments passed to render_template.}
}
\value{
stamp_preview object.
}
\description{
Preview stamp without modifying
}
\keyword{internal}
