% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{preview_update}
\alias{preview_update}
\title{Preview header update}
\usage{
preview_update(file, header)
}
\arguments{
\item{file}{Character. Path to file.}

\item{header}{List. Header information from extract_header.}
}
\value{
stamp_update_preview object.
}
\description{
Preview header update
}
\keyword{internal}
