% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.stamp_dir_results}
\alias{print.stamp_dir_results}
\title{Print method for directory results}
\usage{
\method{print}{stamp_dir_results}(x, ...)
}
\arguments{
\item{x}{stamp_dir_results object.}

\item{...}{Additional arguments.}
}
\value{
The stamp_dir_results object, invisibly.
}
\description{
Print method for directory results
}
