% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.stamp_edits}
\alias{print.stamp_edits}
\title{Print method for header edits}
\usage{
\method{print}{stamp_edits}(x, ...)
}
\arguments{
\item{x}{stamp_edits object.}

\item{...}{Additional arguments.}
}
\value{
The stamp_edits object, invisibly.
}
\description{
Print method for header edits
}
