% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.stamp_file_info}
\alias{print.stamp_file_info}
\title{Print method for file info}
\usage{
\method{print}{stamp_file_info}(x, ...)
}
\arguments{
\item{x}{stamp_file_info object.}

\item{...}{Additional arguments.}
}
\value{
The stamp_file_info object, invisibly.
}
\description{
Print method for file info
}
