% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.stamp_language}
\alias{print.stamp_language}
\title{Print method for language}
\usage{
\method{print}{stamp_language}(x, ...)
}
\arguments{
\item{x}{stamp_language object.}

\item{...}{Additional arguments.}
}
\value{
The stamp_language object, invisibly.
}
\description{
Print method for language
}
