% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.stamp_preview}
\alias{print.stamp_preview}
\title{Print method for stamp preview}
\usage{
\method{print}{stamp_preview}(x, ...)
}
\arguments{
\item{x}{stamp_preview object.}

\item{...}{Additional arguments.}
}
\value{
The stamp_preview object, invisibly.
}
\description{
Print method for stamp preview
}
