% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.stamp_template}
\alias{print.stamp_template}
\title{Print method for templates}
\usage{
\method{print}{stamp_template}(x, ...)
}
\arguments{
\item{x}{stamp_template object.}

\item{...}{Additional arguments.}
}
\value{
The stamp_template object, invisibly.
}
\description{
Print method for templates
}
