% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{print.stamp_update_preview}
\alias{print.stamp_update_preview}
\title{Print method for update preview}
\usage{
\method{print}{stamp_update_preview}(x, ...)
}
\arguments{
\item{x}{stamp_update_preview object.}

\item{...}{Additional arguments.}
}
\value{
The stamp_update_preview object, invisibly.
}
\description{
Print method for update preview
}
