% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_escape}
\alias{regex_escape}
\title{Escape regular expression metacharacters}
\usage{
regex_escape(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with metacharacters escaped.
}
\description{
Escape regular expression metacharacters
}
\keyword{internal}
