% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-template.R
\name{render_template}
\alias{render_template}
\title{Render template with variables}
\usage{
render_template(template, file, ...)
}
\arguments{
\item{template}{stamp_template object. Template to render.}

\item{file}{Character. Path to file.}

\item{...}{Additional variables.}
}
\value{
Character. Rendered template.
}
\description{
Render template with variables
}
\keyword{internal}
