% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-file.R
\name{stamp_file}
\alias{stamp_file}
\title{Stamp a single file with a header}
\usage{
stamp_file(file, template = NULL, action = "modify", ...)
}
\arguments{
\item{file}{Character. Path to file to stamp.}

\item{template}{Character or stamp_template object. Template to use for stamping.}

\item{action}{Character. Action to perform: "modify", "dryrun", or "backup".}

\item{...}{Additional arguments passed to render_template.}
}
\value{
Result object, invisibly.
}
\description{
Stamp a single file with a header
}
