% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-template.R
\name{stamp_template_content}
\alias{stamp_template_content}
\title{Create template content with multiline support}
\usage{
stamp_template_content(...)
}
\arguments{
\item{...}{Character strings to be concatenated.}
}
\value{
stamp_template_content object.
}
\description{
Create template content with multiline support
}
