% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-template.R
\name{stamp_template_create}
\alias{stamp_template_create}
\title{Create a new template}
\usage{
stamp_template_create(name, fields = NULL, content = NULL)
}
\arguments{
\item{name}{Character. Name of template.}

\item{fields}{stamp_template_fields object. Template fields.}

\item{content}{stamp_template_content object. Template content.}
}
\value{
stamp_template object.
}
\description{
Create a new template
}
