% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-template.R
\name{stamp_template_describe}
\alias{stamp_template_describe}
\title{Define template fields}
\usage{
stamp_template_describe(...)
}
\arguments{
\item{...}{Named list of stamp_template_field objects.}
}
\value{
stamp_template_fields object.
}
\description{
Define template fields
}
