% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-template.R
\name{stamp_template_field}
\alias{stamp_template_field}
\title{Define individual field}
\usage{
stamp_template_field(name, default = NULL, required = FALSE)
}
\arguments{
\item{name}{Character. Field name.}

\item{default}{Any. Default value of field.}

\item{required}{Logical. Whether field is required.}
}
\value{
stamp_template_field object.
}
\description{
Define individual field
}
