% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-template.R
\name{stamp_template_load}
\alias{stamp_template_load}
\title{Load template from YAML}
\usage{
stamp_template_load(name)
}
\arguments{
\item{name}{Character. Name of template.}
}
\value{
stamp_template object.
}
\description{
Load template from YAML
}
