% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{stamp_update_helper_author_add}
\alias{stamp_update_helper_author_add}
\title{Helper for adding authors}
\usage{
stamp_update_helper_author_add(new_author)
}
\arguments{
\item{new_author}{Character. New author to add.}
}
\value{
Function to update author field.
}
\description{
Helper for adding authors
}
