% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{stamp_update_helper_copyright_extend}
\alias{stamp_update_helper_copyright_extend}
\title{Helper for extending copyright years}
\usage{
stamp_update_helper_copyright_extend(initial_year = NULL)
}
\arguments{
\item{initial_year}{Character or NULL. Initial year to use if no year is found.}
}
\value{
Function to update copyright field.
}
\description{
Helper for extending copyright years
}
