% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-variables.R
\name{stamp_variables_add}
\alias{stamp_variables_add}
\title{Add custom variable}
\usage{
stamp_variables_add(name, value)
}
\arguments{
\item{name}{Character. Variable name.}

\item{value}{Any. Variable value or function.}
}
\value{
Environment with all variables.
}
\description{
Add custom variable
}
