% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-variables.R
\name{stamp_variables_list}
\alias{stamp_variables_list}
\title{List all available variables}
\usage{
stamp_variables_list()
}
\value{
Character vector of variable names.
}
\description{
List all available variables
}
