% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{update_header_field}
\alias{update_header_field}
\title{Update a header field}
\usage{
update_header_field(header, field, update)
}
\arguments{
\item{header}{List. Header information from extract_header.}

\item{field}{Character. Field name to update.}

\item{update}{Function or character. Update to apply.}
}
\value{
Updated header list.
}
\description{
Update a header field
}
\keyword{internal}
