% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{update_one_file}
\alias{update_one_file}
\title{Apply header updates to a single file}
\usage{
update_one_file(file, updates, action)
}
\arguments{
\item{file}{Character. Path to file.}

\item{updates}{Named list of updates.}

\item{action}{Character. Action to perform.}
}
\value{
TRUE invisibly, FALSE if skipped, or a preview object for "dryrun".
}
\description{
Apply header updates to a single file
}
\keyword{internal}
