% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yaml_front_matter_end}
\alias{yaml_front_matter_end}
\title{Find the end of a YAML front matter block}
\usage{
yaml_front_matter_end(content)
}
\arguments{
\item{content}{Character vector. File content lines.}
}
\value{
Integer. Line number of the closing fence, or 0 if the file does
not start with YAML front matter.
}
\description{
Find the end of a YAML front matter block
}
\keyword{internal}
