% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{extract_header}
\alias{extract_header}
\title{Extract header from file}
\usage{
extract_header(file)
}
\arguments{
\item{file}{Character. Path to file.}
}
\value{
List with header information.
}
\description{
Extract header from file
}
\keyword{internal}
