% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{has_header}
\alias{has_header}
\title{Check if file has a header}
\usage{
has_header(file)
}
\arguments{
\item{file}{Character. Path to file.}
}
\value{
Logical. TRUE if file has a header.
}
\description{
Check if file has a header
}
