% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{stamp_update}
\alias{stamp_update}
\title{Update existing file headers}
\usage{
stamp_update(file, updates, action = "modify")
}
\arguments{
\item{file}{Character. Path to file to update.}

\item{updates}{Named list. Updates to apply to header fields.}

\item{action}{Character. Action to perform: "modify", "dryrun", or "backup".}
}
\value{
TRUE invisibly on success.
}
\description{
Update existing file headers
}
