% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-update.R
\name{update_file_header}
\alias{update_file_header}
\title{Update file header}
\usage{
update_file_header(file, header)
}
\arguments{
\item{file}{Character. Path to file.}

\item{header}{List. Header information from extract_header.}
}
\value{
TRUE invisibly on success.
}
\description{
Update file header
}
\keyword{internal}
